Removed the Completed Plan Fragment
