+++ app/src/main/res/layout/activity_plan_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.Plan.CompletedPlanFragment">
+    tools:context=".ui.Plan.PlanFragment">
 
     <TextView
         android:id="@+id/plan_title"
